Extract and export the `enum` serializer
