+++ packages/umi-serializer-beet/test/enum.test.ts
@@ -1,6 +1,7 @@
 import test from 'ava';
 import { createBeetSerializer } from '../src';
 import { s, d } from './_helpers';
+import { scalarEnum } from '../src/scalarEnum';
 
 enum Empty {}
 enum Feedback {
@@ -15,7 +16,7 @@ enum Direction {
 }
 
 test('numerical enum (de)serialization', (t) => {
-  const { enum: scalarEnum, u64 } = createBeetSerializer();
+  const { u64 } = createBeetSerializer();
 
   // Bad.
   s(t, scalarEnum(Feedback), Feedback.BAD, '00');
@@ -52,8 +53,6 @@ test('numerical enum (de)serialization', (t) => {
 });
 
 test('lexical enum (de)serialization', (t) => {
-  const { enum: scalarEnum } = createBeetSerializer();
-
   // Up.
   s(t, scalarEnum(Direction), Direction.UP, '00');
   s(t, scalarEnum(Direction), 'Up' as Direction, '00');
@@ -97,7 +96,7 @@ test('lexical enum (de)serialization', (t) => {
 });
 
 test('description', (t) => {
-  const { enum: scalarEnum, u32 } = createBeetSerializer();
+  const { u32 } = createBeetSerializer();
   t.is(scalarEnum(Empty).description, 'enum(; u8)');
   t.is(scalarEnum(Feedback).description, 'enum(BAD, GOOD; u8)');
   t.is(
@@ -112,7 +111,7 @@ test('description', (t) => {
 });
 
 test('sizes', (t) => {
-  const { enum: scalarEnum, u32 } = createBeetSerializer();
+  const { u32 } = createBeetSerializer();
   t.is(scalarEnum(Empty).fixedSize, 1);
   t.is(scalarEnum(Empty).maxSize, 1);
   t.is(scalarEnum(Feedback).fixedSize, 1);
