+++ packages/umi-serializer-data-view/src/createDataViewSerializer.ts
@@ -3,11 +3,9 @@ import {
   DataEnum,
   DataEnumSerializerOptions,
   DataEnumToSerializerTuple,
-  EnumSerializerOptions,
   fixSerializer,
   mergeBytes,
   none,
-  ScalarEnum,
   Serializer,
   SerializerInterface,
   StringSerializerOptions,
@@ -40,6 +38,7 @@ import {
 } from './numbers';
 import { option } from './option';
 import { publicKey } from './pubkey';
+import { scalarEnum } from './scalarEnum';
 import { set } from './set';
 import { struct } from './struct';
 import { sumSerializerSizes } from './sumSerializerSizes';
@@ -77,67 +76,6 @@ function getTolerantSerializerFactory<
 export function createDataViewSerializer(
   options: DataViewSerializerOptions = {}
 ): SerializerInterface {
-  const enumFn = <T>(
-    constructor: ScalarEnum<T> & {},
-    options: EnumSerializerOptions = {}
-  ): Serializer<T> => {
-    const prefix = options.size ?? u8();
-    const enumKeys = Object.keys(constructor);
-    const enumValues = Object.values(constructor);
-    const isNumericEnum = enumValues.some((v) => typeof v === 'number');
-    const valueDescriptions = enumValues
-      .filter((v) => typeof v === 'string')
-      .join(', ');
-    const minRange = 0;
-    const maxRange = isNumericEnum
-      ? enumValues.length / 2 - 1
-      : enumValues.length - 1;
-    const stringValues: string[] = isNumericEnum
-      ? [...enumKeys]
-      : [...new Set([...enumKeys, ...enumValues])];
-    function assertValidVariant(variant: number | string): void {
-      const isInvalidNumber =
-        typeof variant === 'number' &&
-        (variant < minRange || variant > maxRange);
-      const isInvalidString =
-        typeof variant === 'string' && !stringValues.includes(variant);
-      if (isInvalidNumber || isInvalidString) {
-        throw new DataViewSerializerError(
-          `Invalid enum variant. Got "${variant}", ` +
-            `expected one of [${stringValues.join(', ')}] ` +
-            `or a number between ${minRange} and ${maxRange}`
-        );
-      }
-    }
-    return {
-      description:
-        options.description ??
-        `enum(${valueDescriptions}; ${prefix.description})`,
-      fixedSize: prefix.fixedSize,
-      maxSize: prefix.maxSize,
-      serialize: (value: T) => {
-        assertValidVariant(value as string | number);
-        if (typeof value === 'number') return prefix.serialize(value);
-        const valueIndex = enumValues.indexOf(value);
-        if (valueIndex >= 0) return prefix.serialize(valueIndex);
-        return prefix.serialize(enumKeys.indexOf(value as string));
-      },
-      deserialize: (bytes: Uint8Array, offset = 0) => {
-        if (bytes.slice(offset).length === 0) {
-          throw new DeserializingEmptyBufferError('enum');
-        }
-        const [value, newOffset] = prefix.deserialize(bytes, offset);
-        const valueAsNumber = Number(value);
-        offset = newOffset;
-        assertValidVariant(valueAsNumber);
-        return [
-          (isNumericEnum ? valueAsNumber : enumValues[valueAsNumber]) as T,
-          offset,
-        ];
-      },
-    };
-  };
-
   const dataEnum = <T extends DataEnum, U extends T = T>(
     variants: DataEnumToSerializerTuple<T, U>,
     options: DataEnumSerializerOptions = {}
@@ -298,7 +236,7 @@ export function createDataViewSerializer(
       ? getTolerantSerializerFactory(nullable, () => null)
       : nullable,
     struct,
-    enum: enumFn,
+    enum: scalarEnum,
     dataEnum,
     string,
     bool,
