+++ packages/umi-serializer-data-view/src/index.ts
@@ -5,6 +5,7 @@ export * from './errors';
 export * from './map';
 export * from './nullable';
 export * from './option';
+export * from './scalarEnum';
 export * from './set';
 export * from './struct';
 export * from './tuple';
